implement proper MVVM
