+++ CupCakeCorner PROJECT/CupCakeCorner PROJECT/ContentView.swift
@@ -9,39 +9,37 @@ import SwiftUI
 
 struct ContentView: View {
     
-    @State private var order = Order()
+    @State private var viewModel = OrderVM()
     
     var body: some View {
         NavigationStack {
             Form {
                 Section {
-                    Picker("Select your cake type", selection: $order.type) {
+                    Picker("Select your cake type", selection: $viewModel.order.type) {
                         ForEach(Order.types.indices, id: \.self) {
                             Text(Order.types[$0])
                         }
                     }
                     
-                    Stepper("Number of cakes: \(order.quantity)", value: $order.quantity, in: 3...20)
+                    Stepper("Number of cakes: \(viewModel.order.quantity)", value: $viewModel.order.quantity, in: 3...20)
                 }
                 
                 Section {
-                    Toggle("Any special requests?", isOn: $order.specialRequestEnabled.animation())
+                    Toggle("Any special requests?", isOn: $viewModel.order.specialRequestEnabled.animation())
                     
-                    if order.specialRequestEnabled {
-                        Toggle("Add extra frosting", isOn: $order.extraFrosting)
-                        Toggle("Add extra sprnkles", isOn: $order.addSprinkles)
+                    if viewModel.order.specialRequestEnabled {
+                        Toggle("Add extra frosting", isOn: $viewModel.order.extraFrosting)
+                        Toggle("Add extra sprnkles", isOn: $viewModel.order.addSprinkles)
                     }
                 }
                 
                 Section {
                     NavigationLink("Delivery Details") {
-                        AddressView(order: order)
+                        AddressView(viewModel: viewModel)
                     }
                 }
             }
-            .navigationTitle("Cupcake Corner")
         }
-       
     }
 }
