--- CupCakeCorner PROJECT/CupCakeCorner PROJECT/Order.swift
@@ -1,108 +1,0 @@
-//
-//  Order.swift
-//  CupCakeCorner PROJECT
-//
-//  Created by Balázs Képli on 6/2/25.
-//
-
-import Foundation
-
-@Observable
-class Order: Codable {
-    
-    enum CodingKeys: String, CodingKey {
-        case _type = "type"
-        case _quantity = "quantity"
-        case _specialRequestEnabled = "specialRequestEnabled"
-        case _extraFrosting = "extraFrosting"
-        case _addSprinkles = "addSprinkles"
-        case _name = "name"
-        case _city = "city"
-        case _streetAddress = "streetAddress"
-        case _zip = "zip"
-        
-    }
-    
-    static let types = ["Vanilla", "Strawberry", "Chocolate", "Rainbow"]
-    
-    var type = 0
-    var quantity = 3
-    
-    var specialRequestEnabled = false {
-        didSet {
-            if specialRequestEnabled == false {
-                extraFrosting = false
-                addSprinkles = false
-            }
-        }
-    }
-    var extraFrosting = false
-    var addSprinkles = false
-    
-    var name = ""
-    var streetAddress = ""
-    var city = ""
-    var zip = ""
-    
-    var confirmationMessage: String = ""
-    var showingConfirmation: Bool = false
-    
-    var hasValidAddress: Bool {
-        if name.isEmpty || streetAddress.isEmpty || city.isEmpty || zip.isEmpty {
-            return false
-        }
-        
-        return true
-    }
-    
-    var cost: Decimal {
-        // 2$ per cake
-        var cost = Decimal(quantity * 2)
-        
-        // complicated cakes cost more
-        cost += Decimal(type) / 2
-        
-        // $1/cake for extra frosting
-        if extraFrosting {
-            cost += Decimal(quantity)
-        }
-        
-        // 0.50$ / cake for extra sprinkle
-        if addSprinkles {
-            cost += Decimal(quantity) / 2
-        }
-        
-        return cost
-    }
-    
-    func placeOrder(for order: Order) async {
-        
-        // Encode order into a JSON
-        guard let encoded = try? JSONEncoder().encode(order) else {
-            print("Failed to encode order")
-            return
-        }
-        
-        // Create url and create a request
-        let url = URL(string: "https://reqres.in/api/cupcakes")!
-        var request = URLRequest(url: url)
-        
-        //Filling the request specs
-        request.setValue("application/json", forHTTPHeaderField: "Content-Type")
-        request.httpMethod = "POST"
-        
-        do {
-            let (data, _) = try await URLSession.shared.upload(for: request, from: encoded)
-            
-            print(data)
-            
-            
-            // handle the result, confirm that resposne is equal to request json
-            let decodedOrder = try JSONDecoder().decode(Order.self, from: data)
-            confirmationMessage = "Your order for \(decodedOrder.quantity) x\(Order.types[decodedOrder.type].lowercased()) cupcakes is on the way!"
-            showingConfirmation = true
-        } catch {
-            print("Check out failed: \(error.localizedDescription)")
-        }
-    }
-}
