+++ CupCakeCorner PROJECT/CupCakeCorner PROJECT/OrderModel.swift
@@ -1,0 +1,77 @@
+//
+//  OrderModel.swift
+//  CupCakeCorner PROJECT
+//
+//  Created by Balázs Képli on 21/2/25.
+//
+
+import Foundation
+
+struct Order: Codable, Hashable, Identifiable {
+    
+//    enum CodingKeys: String, CodingKey {
+//        case _id = "id"
+//        case _type = "type"
+//        case _quantity = "quantity"
+//        case _specialRequestEnabled = "specialRequestEnabled"
+//        case _extraFrosting = "extraFrosting"
+//        case _addSprinkles = "addSprinkles"
+//        case _name = "name"
+//        case _city = "city"
+//        case _streetAddress = "streetAddress"
+//        case _zip = "zip"
+//    }
+//        
+        
+    var id = UUID()
+    
+    static let types = ["Vanilla", "Strawberry", "Chocolate", "Rainbow"]
+    
+    var type = 0
+    var quantity = 3
+    
+    var specialRequestEnabled = false {
+        didSet {
+            if specialRequestEnabled == false {
+                extraFrosting = false
+                addSprinkles = false
+            }
+        }
+    }
+    
+    var extraFrosting = false
+    var addSprinkles = false
+    
+    var name = ""
+    var streetAddress = ""
+    var city = ""
+    var zip = ""
+    
+    var hasValidAddress: Bool {
+        if name.isEmpty || streetAddress.isEmpty || city.isEmpty || zip.isEmpty {
+            return false
+        }
+        
+        return true
+    }
+    
+    var cost: Decimal {
+        // 2$ per cake
+        var cost = Decimal(quantity * 2)
+        
+        // complicated cakes cost more
+        cost += Decimal(type) / 2
+        
+        // $1/cake for extra frosting
+        if extraFrosting {
+            cost += Decimal(quantity)
+        }
+        
+        // 0.50$ / cake for extra sprinkle
+        if addSprinkles {
+            cost += Decimal(quantity) / 2
+        }
+        
+        return cost
+    }
+}
